+++ holberton.h
@@ -4,7 +4,7 @@
 int _printf(const char *format, ...);
 int _putchar(char c);
 int print_str(char *c);
-int print_int(long int num);
+int print_int(int num);
 int print_Str(char *c);
 int print_bin(unsigned int num);
 int print_hex(unsigned int num);
